requirements.txt
Django==5.2.1
django-storages==1.14.6
google-cloud-storage==3.1.0
google-auth==2.40.2
google-auth-oauthlib==1.2.2
opencv-python-headless==4.10.0.84
pytesseract==0.3.13
Pillow==11.0.0
psycopg2-binary==2.9.10
easyocr==1.7.2
openai==1.3.8
httpx==0.25.2
python-dotenv==1.1.0
gunicorn==22.0.0
whitenoise==6.8.2
requests==2.31.0
transformers==4.46.3
torchaudio==2.5.1
ffmpeg-python==0.2.0
accelerate==1.2.1 